Updates to support Android 12 Fixed error when setting background image Removed additional unnecessary legacy code Updated libraries

+++ gic_flutter/lib/src/views/screenEditor/backgroundDialog.dart
@@ -62,7 +62,8 @@ class _BackgroundDialogState extends State<BackgroundDialog> {
       Directory dest = await getApplicationDocumentsDirectory();
       String destPath = path.join(
           dest.path, "screens", screenViewModel.screenId.toString(), path.basename(file.path));
-      File(screenViewModel.backgroundPath).deleteSync();
+      if (File(screenViewModel.backgroundPath).existsSync())
+        File(screenViewModel.backgroundPath).deleteSync();
       File newFile = File(file.path).copySync(destPath);
       setState(() {
         screenViewModel.backgroundPath = newFile.path;
